Payment package classes were refactored a little bit
