+++ src/main/java/com/zufar/onlinestore/payment/service/impl/PaymentServiceImpl.java
@@ -3,7 +3,7 @@ package com.zufar.onlinestore.payment.service.impl;
 import com.stripe.exception.StripeException;
 import com.zufar.onlinestore.payment.dto.PaymentDetailsWithTokenDto;
 import com.zufar.onlinestore.payment.dto.PaymentDetailsDto;
-import com.zufar.onlinestore.payment.model.Payment;
+import com.zufar.onlinestore.payment.entity.Payment;
 import com.zufar.onlinestore.payment.exception.PaymentNotFoundException;
 import com.zufar.onlinestore.payment.mapper.PaymentConverter;
 import com.zufar.onlinestore.payment.processor.PaymentProcessor;
